Update 1 - Player class, Movement and animation attempt
Added player class and movement controls. Tried to add animation for walking (in the move subroutine for the player class) but am struggling to get it working due to each sprite in the animation being different sizes. Approached animation by cycling through an array of images when the character is walking