Update 4 - Menu and background image
I added a menu screen that the game starts on with the option to choose the multiplayer game mode (will add more options later). Also added background for the first map. Atm the player cannot choose the map but i will add a screen to do that once i begin to create the maps. Had difficulty centering the text for the menu screen, opted for coordinates that looked about right but will try to centre the text properly later. Currently have a glitch where the player no longer moves smoothly across the screen, and only moves each time the key is pressed down.