Update 5 - Bug fixes
Fixed bug where the player only moved when a key was pressed down. The indentation was incorrect so that the multiplayer code only ran when there was an event in pygame, so only when a key was pressed down. Also fixed the white background square around the sprites. Took a long time learning to photoshop images to make the background transparent, and also found out that the transparency only carried over if i did not convert the images.