Update 6 - Falling and floors
Added a hard floor class (A floor you will not be be able to jump up through) and made the character fall, accelerating at 9.8 pixels a second. The character then stops falling and stands on the floor when he hits it. Changed some sprites heights so they are all the same.