Update 7 - Crouching
Added the ability for the player to crouch. Had problem where since the player sprite is now smaller it was not alligned with the ground properly. Used the self.crouching and uncrouching attributes to adjust the coordinates of the player on the first frame he is crouched and the first where he stands up. Also made not being crouched a condition of walking and not being in the air a condition of crouching, so that the player cannot moved whilst crouched or crouch in the air to make themselves smaller