Update 8 - Shooting animations
When the player presses n, the chracter sprite switches tot he shooting image. This was simple, but making this work whilst the character was crouching was tricky as the sprite coords had to be adjusted. Line 242 to 254 and 270 to 278 fix this by remembering if the character was crouching with the wascrouched variable and setting the self.crouching and uncrouching variable to get the programme to readjust the sprites coords when appropriate