Update 9 - Shooting and Player 2
Added shooting mechanics, but not the aiming ones, the player can only fire forward or backwards. Bullets now despawn when they collide with any obstacles. Added the second player, need to add the health and collisions with bullets and players. No real bugs - forgot to change the key down from n to 2 for player 2 but fixed this quickly