Update 10 - Player and bullet collisions
Added the collisions between the player and the bullet in the player update function. Gave the player 100 health and subtract 9 each time he is hit. Had a problem where the bullet was colliding with the player shooting it immediately, so adjusted the coordinates where the bullets are spawned a bit (line 384 and 311).