Update 15 - Back to Menu when game ends
I added a Gameover variable which is set to true when 1 players health drops to 0. This then causes Multiplayer to stop running and Gameover to start running. Gameover is a screen saying the winner and to press space to return to the menu. The User can press space, return to the menu and start a new game. I had a bug where the player could not restart the game without it ending immediately and the gameover screen being displayed, declaring the winner as the winner of the last game. This was due to the player sprites stats not being reset so 1 of their health was already bellow 0. I fixed this by killing all sprites when Gameover is set to true and re-initiating the players when the new game is started