Update 16 - Soft Floors
I tried to add floors that the player can jump up through, but then stand on once on top. It was tricky to find the right height, but after playing around with the jump force, gravity and height of the platforms i found that a height of 100 pixels above the platform bellow works well with a -5.4 jump power and 0.2 acceleration. However, it is difficult to screen shot but I have a bug where the player will teleport to be standing on the platform as long as he touches it and is moving downwards. This is a problem because only his head may be touching it but he will be placed on top. Need to fix this next