Update 17 - Fixed soft floor glitch
I fixed the glitch (from the previous update) where the character would teleport up if he made contact with the floor at all. I did this by adding the algorithm on lines 192 - 207. What this does is essentially:
1. checks if the player is in contact with a floor and falling
2. If so, checks if the bottom of the character is less than 5 pixels (his head is more that 58 above) below the top of the platform. 
3. If it is, he is deemed high enough the "land" on the platform and is placed on top of it.

However, I had a glitch where the player started phasing through floors if he fell from a certain height. After a while, I realised that this was happening because the character was accelerating to a point where they were phasing through this 5 pixel window to land on the platform in 1 frame. To fix this, I added the second part of the or statement on line 200, so that if the character was falling at more than 5 pixels a frame, his window for landing was extended to 15 pixels. This should be more than enough since the screen is only so big so the character cannot speed up too much.

Note: I figured out 15 pixels by taking speed to = t(0.2)     t in frames
therefore displacement = 0.1t^2.     I took the maximum screen size to be 550 (750-100 for stats at top - 100 for hardfloor at bottom) and so solved for t. i then plugged this back into the speed equation to get just under 5
This can be improved if necessary by adding more speed ranges such as 5 and above with 1 window, 10 and above with another.