Update 18 - Finished Map 1
Finished creating the floors for map 1. Added the gun spawns In the right locations. No difficulties here. Screenshot in the screenshot file of map 1 at this point. Next I need to add different statistics for the magnum e.g. different damage and add a cool down after each shot to prevent the player firing as fast as they can pull the trigger.