Update 19 - Added different damage for different guns
Small update. Added that the magnum and the pistol do different damage. Took along time since i had a glitch where i put the bullet.gun = player1/2.gun in the wrong place, so that it only ran if the direction of the bullet was right. Once this was spotted it was easy to fix and all worked well