Update 20 - Ammo for guns
Added that guns have limitted ammo. Pistol has 12, magnum has 5 and niether gun can shoot when out. Updated the draw stats function to display this. No problems here. Also removed a faulty line of code in the player2 key up for the 2 key which was effecting player1's state. I hadn't noticed it before and it didn't seem to be causing any issues