Update 22 - Fixed floor glitch
The floor seems to spawn in correctly when initiated in the main body of code instead of  through a subroutine. Not sure why but keeps happening so I've gone with it and moved the level create code to the main body - it'll be the same length in lines anyway. Also have a bug where bullets don't move properly in single player mode, they just move at the same speed as everything else so I need to fix that next