Update 23 - Single Player Level 1 Map
Added the code to generate the map for level 1. The map design is very similar to map 1 from multiplayer, but I added another floor in the middle row in between each set of  map 1's  so that the map flows fluently into the next set. Currently have an issue where the floors move closer together as they move left across the screen, until they overlap each other. Screenshot in files