Update 24 - Fixed overlapping floors glitch
Found the reason the floors were overlapping as they moved left across screen. The glitch only happened when the sprites reached the very left of the screen. I realised that the problem was that when the sprites top left corner was on the screen, it was travelling faster than when it was not. I think the problem was that because they were being moved 2.5 pixels each frame, the program was moving the top left corner of the sprite 2.5 pixels left each frame and then rounding this down to essentially move it 3 pixels. However, when the top left corner of the floor sprites travelled off the screen, their coordinates went negative, so the program started rounding them down instead of up. I'm not sure why it was doing this, but I changed the amount everything moves each frame to 2 and everything worked as expected