Update 25 - Added enemies to single player
Added enemies to the single player mode. They spawn as expected and move across the screen, but at the moment cannot stand on platforms (they fall straight through) and jump whether they are on a platform or not. Need to change this for next update so that they run along platforms, move slower, spawn less frequently and only jump when on a platform