Update 27 - Enemies now disappear when shot and stats shown at top of screen
Added that when enemies are shot they disappear, along with the bullet. Also added the players stats being drawn at the top of the screen. Had the same glitch as earlier where when the instructions for drawing the stats are in a subroutine they don't work, even though the exact same code works when put into the main code body and the function works when called in the exact same way in the multiplayer game mode. Currently the code is in the main body, but I need to fix this glitch soon .
Also added a player lives stat, so that the player can loose a life for each enemy that gets to the other side.