Update 28 - Single player game ends when lives drop to 0
Made it so that when an enemy moves off the screen the player looses a life. When lives drop to 0, all sprites are deleted, the game shows the gameover screen which shows that the player has lost and instructs them to press space to return to the menu, allowing them to play again. Also added that when a platform moves off the screen it is deleted in order to optimise the game (Added previously but forgot to mention).