Update 29 - Game stores L1 top score
Added a feature so that the game remembers the top score from L1 in a text file. Tested this feature using print(L1topscore) at the beginning of the programme and seeing if the correct top score was printed. Removed this for commit but can be added back for write up/screenshots etc. This was tricky as I originally tried to use pickle and then shelve for this feature, with no success. Once I realised python had an inbuilt feature for accessing text files things went quite well and the programme worked quite quickly