Update 30 - Game displays top score for each level on level select screen
Added that the game displays the top score for each level on the level select screen. Did this by just blitting the text to the screen (see line 480). Two very minor issues here: 1 was that since the L2 and L3 text files had nothing in the score came out as a blank space - fixed this by editing the text files to contain 0 until they are set to something else by the game, 2nd was just playing around with coordinates until they seemed centred. Maybe later on I can centre them properly but for now they are close enough that they appear centred