Update 31 - Fixed high score glitch
Found glitch where if the player beat their high score on a level, and then selected to play single player again the programme would crash as L1topscore was saved as an integer (since its value was copied form score which is an integer) and integers cannot be outputted using .render. Changed line 705 to convert score to a string before passing its value to L1topscore. Made a note to include this string conversion when I add level 2 and 3. Screenshot of the console output for this bug in screenshots