Update 36 - Changed enemy AI
Trying to add to enemy AI so that they are more likely to jump on Level 3 in order to make it harder as it is the final level and so that enemies are more likely to jump on all levels if the player is on the same x coordinates as them i.e. they are able to be aiming up a shot. Currently have a bug where on level 3, all the enemies jump as soon as thye come in contact with a floor.