Update 37 - Fixed enemy AI
Fisxed some glitches with enemy AI changes. Firstly realised a bug with the enemies being more likely to jump when on the same height - i was comparing x coordinates instead of y coords so fixed this. Also changed the way i wrote the if variable to if jump = 1 or jump = 2. This fixed the instant jumping glitch on level 3