Update 43 - Fixed map rendering glitches
Glitches were because when I blitted the background images I had if map == "1" instead of Map. map seemed to be a pre-built function in python or pygame so was causing issues. I'm not sure why the logo or text were still being blitted, but everything worked once i fixed this typo so I presume it was something to do with the pre-built function