Update 52 - fixed sprite not appearing when cimbing glitch
The climb sprite not showing up was ue to me refering to the players stae as "climbing" in the main body and "climb" in the update fiunction so i fixed this. The dimensions of the climb sprite (50, 63) were also far to wide so i adjusted this to 35,63. However, the player still does not move up the ladder despite me adding code that says they are supported when they are climbing and so should not fall down due to gravity so I need to fix this gltich next