Update 53 - Fixed map climbing glitch
Fixed the glitch where the player does not climb the ladder. This is because I am a fool and put the code to stop the player falling with gravity when climbing a ladder int he enemy update function rather than the player 1. However, I have a new bug where the player does not fall if they are touching a platform after climbing a ladder, even if just their head is touching it. Screenshot in folder.