Update 59 - Moving platforms class update function
Added some code to the movingsoftfloor init functions to ensure that parameter1 is the larger of the two (i.e. the bottom or the right of the screen). This is so that I can be confident which parameter is larger when I reference them in the update subroutine. I also made the update subroutine, although for now it only covers movement in the y direction. Next I need to test the subroutine (and add the code to call it for each platform).