Update 87 - tried to get player to move with floor
Update 87 - Tried to get player to move along with floor when standing on it
It's not working atm. I used a print(hi) command to find that this is because the whole blaock of code isn't running. I think this is due to the if statment reading if floor == Movingsoftfloor which doesn't work, so i need to find another way of identifying the type of floor the player is standing on