Update 90?
Moving floors now work. The problem was that when the player stood on top of a moving floor, they were not actually colliding with it each frame, they had to wait to fall 1 pixel, which takes a few frames due to the speed being so low at the start of a fall, when they would collide with the moving floor, be put back above it and moved the 2 pixels along with it, but only in this frame. I also realised that this was how the code worked for the still floors: the player would be put above the platform, and then slowly start to fall down through it, every few frames it would fall 1 pixel, and then the code would immediately place it back above the platform, so the user never actually saw the player fall down. To fix the glitch i added a line of code so that if the floor the player collides with is a moving floor, they are placed 1 pixel further down, so that they are colliding woth the floor at all times and hence are moved along the 2 pixels a frame. I also moved the code that does this into its own section above the code that places the player above the floors, so that it can be ran every frame, regardless of if the players y speed is 0 or not